+++ packages/commands/src/index.ts
@@ -1,2 +1,2 @@
 export * from './Command'
-export * from './Service'
+export * from './CommandService'