+++ packages/http/src/content-type/index.ts
@@ -1,2 +1,0 @@
-export { HttpError } from './http_error'
-export { HttpJson } from './http_json'