+++ packages/seed/seed/something.ts
@@ -1,0 +1,6 @@
+export default {
+  expiresInDays: 0,
+  hook: (fixture: any) => {
+    return 'qwer'
+  },
+}